requirements.txt
chardet==3.0.4
Django==1.11.4
django-oauth-toolkit==1.0.0
djangorestframework==3.6.4
oauthlib==2.0.2
requests==2.18.4
urllib3==1.22
django-filter==1.0.2
psycopg2==2.7.4
